fix: jest spyOn was not working with classes decorated with @AutoLog
jest-mock `spyOn` is not able to spy a Proxy class that has a `get` trap
that returns a wrapped function.

The fix here is pretty horrible as it makes it so autoreggol knows about
jest's inner workings but it is necessary for now to avoid breaking
people's tests.

This commit also adds:
- Jest to be able to test the issue itself
- Eslint